[ARO-17938] Check Gateway VMSS Instance Health Before Stopping Gateway Service (#4269)
* kevinobriendotca/ARO-17938-Check-Gateway-VMSS-Instance-Health-Before-Stopping
- Added health check before attempting to run gateway service stop command on gateway VM.
- Fixed logic bug in isVMInstanceHealthy: OR > AND to catch nils along the way to the actual health status code.
- Added wait timeout for errors channel in gatewayRemoveOldScaleset.

* kevinobriendotca/ARO-17938-Check-Gateway-VMSS-Instance-Health-Before-Stopping
- Removed commented code.

* kevinobriendotca/ARO-17938-Check-Gateway-VMSS-Instance-Health-Before-Stopping
- Added test for nil VMHealth field in VirtualMachineScaleSetVMInstanceView.

+++ pkg/deploy/upgrade_gateway.go
@@ -80,19 +80,26 @@ func (d *deployer) gatewayRemoveOldScaleset(ctx context.Context, vmssName string
 	d.log.Printf("stopping scaleset %s", vmssName)
 	errors := make(chan error, len(scalesetVMs))
 	for _, vm := range scalesetVMs {
-		go func(id string) {
-			errors <- d.vmssvms.RunCommandAndWait(ctx, d.config.GatewayResourceGroupName, vmssName, id, mgmtcompute.RunCommandInput{
-				CommandID: to.StringPtr("RunShellScript"),
-				Script:    &[]string{"systemctl stop aro-gateway"},
-			})
-		}(*vm.InstanceID) // https://golang.org/doc/faq#closures_and_goroutines
+		if d.isVMInstanceHealthy(ctx, d.config.GatewayResourceGroupName, vmssName, *vm.InstanceID) {
+			d.log.Printf("stopping gateway service on %s", *vm.Name)
+			go func(id string) {
+				errors <- d.vmssvms.RunCommandAndWait(ctx, d.config.GatewayResourceGroupName, vmssName, id, mgmtcompute.RunCommandInput{
+					CommandID: to.StringPtr("RunShellScript"),
+					Script:    &[]string{"systemctl stop aro-gateway"},
+				})
+			}(*vm.InstanceID) // https://golang.org/doc/faq#closures_and_goroutines
+		}
 	}
 
 	d.log.Print("waiting for instances to stop")
 	for range scalesetVMs {
-		err := <-errors
-		if err != nil {
-			return err
+		select {
+		case err := <-errors:
+			if err != nil {
+				return err
+			}
+		case <-time.After(10 * time.Second):
+			continue
 		}
 	}
